requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.8
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
sentence-transformers==2.2.2
beautifulsoup4==4.12.2
requests==2.31.0
duckduckgo-search==3.9.6
python-multipart==0.0.6
reportlab==4.0.7
python-dotenv==1.0.0
lxml==4.9.3
httpx==0.25.2
anyio==4.0.0
psutil==5.9.6
openpyxl==3.1.2